pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sarma-launcher"
version = "0.1.0"
description = "Distributed Sarma Launcher for coding tasks across multiple Dev Boxes"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "redis>=5.0",
    "pydantic>=2.0",
    "requests>=2.31",
    "azure-identity>=1.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "fakeredis>=2.20",
]

[project.scripts]
sarma = "sarma.master.cli:cli"
sarma-worker = "sarma.worker.runner:main"

[tool.setuptools.packages.find]
where = ["."]
